+++ libgeo/rmxn_shift.c
@@ -1,5 +1,5 @@
 /* See rmxn_shift.h. */
-/* Last edited on 2024-11-22 05:37:41 by stolfi */
+/* Last edited on 2024-11-23 18:54:49 by stolfi */
 
 #include <stdio.h>
 #include <stdint.h>
@@ -17,16 +17,16 @@
 #include <rmxn_shift.h>
 
 void rmxn_shift_rows(uint32_t m, uint32_t n, double A[], double v[], double M[])
-  { for (int32_t i = 0; i < m; i++)
-      { for (int32_t j = 0; j < n; j++) 
-          { int32_t ij = i*(int32_t)n + j; M[ij] = A[ij] + v[j]; }
+  { for (uint32_t i = 0;  i < m; i++)
+      { for (uint32_t j = 0;  j < n; j++) 
+          { uint32_t ij = i*n + j; M[ij] = A[ij] + v[j]; }
       }
   }
   
 void rmxn_shift_cols(uint32_t m, uint32_t n, double v[], double A[], double M[])
-  { for (int32_t i = 0; i < m; i++)
-      { for (int32_t j = 0; j < n; j++) 
-          { int32_t ij = i*(int32_t)n + j; M[ij] = A[ij] + v[i]; }
+  { for (uint32_t i = 0;  i < m; i++)
+      { for (uint32_t j = 0;  j < n; j++) 
+          { uint32_t ij = i*n + j; M[ij] = A[ij] + v[i]; }
       }
   }
